# backend/api/openapi.gen.json
{
  "components": {
...
    "parameters": {
...
      "TenantSlug": {
        "description": "The tenant's slug (docs/adr/0005 D4)",
        "in": "path",
        "name": "tenant",
        "required": true,
        "schema": {
          "pattern": "^[a-z0-9][a-z0-9-]{1,62}$",
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Problem": {
        "content": {
          "application/problem+json": {
            "example": {
              "code": "not_found",
              "detail": "no such tenant",
              "instance": "/api/v1/tenants/acme",
              "request_id": "0199a3c2-1d2e-7f00-8000-0000000000ff",
              "status": 404,
              "title": "Not found",
              "type": "https://cowork.dev/problems/not-found"
            },
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "An RFC 9457 problem details body with a stable `code`, the `request_id`\nand, for validation failures, `errors[]` (docs/adr/0047).\n",
        "headers": {
          "X-Request-Id": {
            "description": "The request's id, also in the body and in the log line",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "schemas": {
...
      "FieldError": {
        "properties": {
          "current": {
            "description": "The server's current value of the field, on a 412; null when the field is empty",
            "type": [
              "string",
              "number",
              "integer",
              "boolean",
              "object",
              "array",
              "null"
            ],
            "x-go-type": "interface{}"
          },
          "message": {
            "type": "string"
          },
          "pointer": {
            "description": "A JSON pointer into the body, or query:\u003cname\u003e / header:\u003cname\u003e",
            "type": "string"
          }
        },
        "required": [
          "pointer",
          "message"
        ],
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "instance": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object"
      },
      "ProblemCode": {
        "description": "The stable code of a problem (docs/adr/0047 D2)",
        "enum": [
          "validation_failed",
          "idempotency_key_required",
          "invalid_cursor",
          "page_too_deep",
          "unauthenticated",
          "token_expired",
          "token_revoked",
          "not_allowed",
          "invalid_credentials",
          "forbidden",
          "insufficient_scope",
          "agent_forbidden",
          "session_required",
          "password_change_required",
          "not_initialised",
          "csrf",
          "not_found",
          "person_not_found",
          "method_not_allowed",
          "username_taken",
          "tenant_slug_taken",
          "project_key_taken",
          "person_ambiguous",
          "grant_exists",
          "mapping_exists",
          "last_admin",
          "project_archived",
          "state_conflict",
          "parent_cycle",
          "link_cycle",
          "open_prerequisites",
          "period_locked",
          "attachment_limit",
          "uploads_disabled",
          "precondition_failed",
          "payload_too_large",
          "unsupported_media_type",
          "idempotency_mismatch",
          "precondition_required",
          "too_many_attempts",
          "internal",
          "not_ready",
          "timeout",
          "backend_unreachable"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "description": "A personal access token, `cwk_` and 43 base62 characters\n(docs/adr/0035 D1). `X-Cowork-Agent: name/model/session` marks the\nrequest as an agent's; a token with the agent flag is an agent's\nwith or without it, and no header removes the flag (docs/adr/0036).\nA request that carries an `Authorization` header is a token's, and\nits cookie, if any, is not looked at. A token of a person the identity\nprovider's groups no longer admit is refused with `401 not_allowed`,\nchecked against the person's groups as of their last login or groups\nrefresh, at most `COWORK_OIDC_GROUPS_REFRESH` apart (docs/adr/0035 D8).\n",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "The session of a browser login: an opaque id of 256 random bits, set\nby `POST /auth/local` or `GET /auth/callback` as `HttpOnly; Secure;\nSameSite=Lax; Path=/` (docs/adr/0031 D2). A session acts as its\nperson with no agent flag and with the whole of the person's role: it\nhas no scope of its own.\nIts writes are held to the CSRF check of docs/adr/0037 D1: `Origin`\n(or `Referer`) equal to the installation's `COWORK_BASE_URL` and\n`X-Requested-With: cowork`.\n",
        "in": "cookie",
        "name": "__Host-cowork-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A multi-tenant backlog for people and the agents that work for them.\n\nEvery route that names a person is authenticated with one of two\ncredentials. A personal access token, `Authorization: Bearer cwk_…`\n(docs/adr/0035), is for scripts and agents; the session cookie of a\nbrowser login (docs/adr/0031 — `POST /auth/local`, or the identity\nprovider's `GET /auth/callback`, docs/adr/0029) is for the UI. Each\noperation says which it takes: most take either, a few — creating a token,\na tenant or a local account, resetting or changing a password, logging out,\nand the administration acts that give access: adding a member, setting a\ngrant, making or changing a group mapping, restricting or opening a\nproject, putting a person on its access list — a session only. A request\nof a session that writes must come from the installation's origin and\ncarry `X-Requested-With: cowork` (docs/adr/0037). A request marked as an\nagent's — by the token's agent flag, or by an `X-Cowork-Agent:\nname/model/session` header — is bound by the agent rules\n(docs/adr/0036, 0043). Errors are RFC 9457 problem details with a stable\n`code` (docs/adr/0047).\n",
    "title": "cowork",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tenants/{tenant}/events": {
      "get": {
        "description": "`ticket.changed` (an upload included), `comment.changed`, `question.changed`,\n`link.changed` and `interest.changed`, each with `data: {\"key\", \"version\", \"kind\"}` — a key\nand a version, never content (docs/adr/0054 D2); `id:` is the act's id. An event of a\nproject the caller cannot see, or of a confidential ticket they may not read, never\narrives (D3).\n\n`membership.changed` says that who belongs to the tenant, or who sees a project, changed —\nby an administrator's act or by the identity provider's groups (docs/adr/0030 D2, D6). Its\n`data` carries keys only, each where it applies: `person_id`, the person whose membership\nor project access changed; `project_id`, the project whose restriction or access list\nchanged; `mapping_id`, the group mapping made, changed or removed. The client reloads what\nit shows of members, mappings and access lists, its project list when `project_id` is\nthere, and the person's own `GET /api/v1/me`. An event about an access entry reaches the\ntenant's administrators and the person it names, one about a mapping the administrators,\nevery other every member.\n\nA reconnect with `Last-Event-ID` inside the replay window replays the gap;\nbeyond it the stream starts with `event: resync`, and the client refetches its lists (D5).\nA comment every twenty seconds keeps proxies open and checks the token and the\nmembership again; the stream ends when either fails. `event: unavailable` ends a stream\nthe server closes — the person's eleventh, a shutdown — and the client polls (D7, D8).\n",
        "operationId": "streamEvents",
        "parameters": [
          {
            "description": "The id of the last event the client received",
            "in": "header",
            "name": "Last-Event-ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The stream"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "The tenant's events the caller may see, as server-sent events",
        "tags": [
          "events"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/TenantSlug"
        }
      ]
    },
...
  },
  "security": [
    {
      "bearerToken": []
    },
    {
      "sessionCookie": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
...
}